[jwies/review-scripted-dispatch] review: split the self-contained glob matcher and dispatch-mode tests out
main's lens-payload seam (#281) left router.ts at 999 lines against the
1000-line cap, so this branch's twelve lines of dispatchMode plumbing
breached it, and the unioned dispatch-mode tests did the same to
router.test.ts. Move the section already marked self-contained into
glob-match.ts (re-exported from router.ts so no importer changes) and the
dispatch-directive CLI tests into router-dispatch-mode.test.ts, the same
split dispatch-contracts.ts took from dispatch.ts.
