+++ workflows/review/lib/router.test.ts
@@ -986,24 +986,3 @@ describe("runCli: lens payload warnings", () => {
         );
     });
 });
-
-describe("runCli: dispatch mode", () => {
-    it("surfaces a configured dispatch scripted line in routing.json", () => {
-        const {fs} = fakeFs({
-            ["/tmp/gh-aw/review/files.json"]: JSON.stringify([
-                {path: "a.ts", status: "modified"},
-            ]),
-            [ROUTING_CONFIG_PATH]: "dispatch scripted",
-        });
-        expect(runCli(fs).dispatchMode).toBe("scripted");
-    });
-
-    it("defaults to task without a dispatch line", () => {
-        const {fs} = fakeFs({
-            ["/tmp/gh-aw/review/files.json"]: JSON.stringify([
-                {path: "a.ts", status: "modified"},
-            ]),
-        });
-        expect(runCli(fs).dispatchMode).toBe("task");
-    });
-});
